plot heatmap
ggplot() is to plot object.

geom_sf() is to plot spatial data. 

data = is to specify the data source. 

aes(fill = `Inventory (Animal heads)`) 
- aes() alllows ggplot() to fill colors of each counties based on values of column Inventory (Animal heads). scale_fill_gradient() then determines how the values are translated into colors. 
- aes() is not used in the second geom_sf() because it is not a data-driven fill; hence, fill = NA as we only want to display the outlines of the states. 
- backticks (``) are used instead of double quotes ("") to tell ggplot2 that Inventory (Animal heads) is a column name, not a string. 

color = is to set the county/state outline color. 

size = is to set the county/state oultline thickness. 

coord_sf() is to set the coordinate reference system of the output plot. 
- The set CRS is Albers Equal Area Conic projection.

labs() is to set plot labels.

them_void() is to remove unnecessary plot elements like gridlines.